requirements.txt
flask
python-dotenv
click
sqlalchemy
werkzeug
Pillow
